+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: dcsicsak <dcsicsak@student.42.fr>          +#+  +:+       +#+         #
+#    By: icseri <icseri@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/11/20 10:41:15 by icseri            #+#    #+#              #
-#    Updated: 2024/12/06 10:00:14 by dcsicsak         ###   ########.fr        #
+#    Updated: 2024/12/06 13:30:36 by icseri           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -15,10 +15,13 @@ NAME = cub3d
 MLX_FLAGS = -lmlx -lXext -lX11 -lm -lz -lbsd
 
 SRCS = main.c \
+		init.c \
 		parsing.c \
-		cleanup.c \
 		parsing_utils.c \
 		parsing_textures.c \
+		parsing_map.c \
+		mlx_utils.c \
+		cleanup.c \
 		cast_rays.c \
 		minimap.c \
 		draw_utils.c \
