+++ mlx_utils.c
@@ -1,0 +1,45 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   mlx_utils.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: icseri <icseri@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/12/06 13:28:30 by icseri            #+#    #+#             */
+/*   Updated: 2024/12/06 13:29:11 by icseri           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+void	render_scene(t_data *data)
+{
+	int	x;
+	int	wall_height;
+
+	wall_height = 0;
+	x = -1;
+	while (++x < WIDTH)
+	{
+		if (data->ray_distance[x] > 0)
+		{
+			wall_height = (TILE_SIZE / data->ray_distance[x]) * (HEIGHT / 55);
+			draw_vertical_line(data, x, wall_height);
+		}
+	}
+}
+
+int	close_window(t_data *data)
+{
+	if (data->img)
+		mlx_destroy_image(data->mlx, data->img);
+	if (data->win)
+		mlx_destroy_window(data->mlx, data->win);
+	if (data->mlx)
+	{
+		mlx_destroy_display(data->mlx);
+		free(data->mlx);
+	}
+	safe_exit(&data->map, EXIT_SUCCESS);
+	return (0);
+}
