+++ parsing_utils.c
@@ -3,41 +3,28 @@
 /*                                                        :::      ::::::::   */
 /*   parsing_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dcsicsak <dcsicsak@student.42.fr>          +#+  +:+       +#+        */
+/*   By: icseri <icseri@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/20 16:26:07 by icseri            #+#    #+#             */
-/*   Updated: 2024/12/05 16:17:15 by dcsicsak         ###   ########.fr       */
+/*   Updated: 2024/12/06 12:59:40 by icseri           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
-void			set_colors(char **colors, int *color, t_map *map);
-unsigned char	get_number(char *str, char **colors, t_map *map);
+void	set_colors(char **colors, int *color, t_map *map);
+int		get_number(char *str, char **colors, t_map *map);
 
 void	get_color(char *line, int *color, t_map *map)
 {
 	char	**color_list;
 
-	if (*color != -1)
-	{
-		ft_free(&line);
-		print_error(1, "Error\nDuplicate color");
-		safe_exit(map, EXIT_FAILURE);
-	}
-	if (ft_strlen(line) <= 2)
-	{
-		ft_free(&line);
-		print_error(1, "Error\nInvalid color");
-		safe_exit(map, EXIT_FAILURE);
-	}
+	if (*color != -1 || ft_strlen(line) <= 2)
+		return (ft_free(&line), safe_exit(map, COLOR));
 	color_list = ft_split(line + 2, ',');
 	ft_free(&line);
 	if (!color_list)
-	{
-		print_error(1, "Error\nMalloc fail");
-		safe_exit(map, EXIT_FAILURE);
-	}
+		safe_exit(map, MALLOC);
 	set_colors(color_list, color, map);
 }
 
@@ -49,17 +36,13 @@ void	create_and_add(char *line, t_list **map_list, t_map *map)
 	content = ft_strtrim(line, "\n");
 	ft_free(&line);
 	if (!content)
-	{
-		free_list(map_list);
-		safe_exit(map, EXIT_FAILURE);
-	}
+		return (free_list(map_list), safe_exit(map, MALLOC));
 	new = ft_lstnew(content);
 	if (!new)
 	{
 		ft_free(&content);
 		free_list(map_list);
-		print_error(1, "Error\nMalloc fail");
-		safe_exit(map, EXIT_FAILURE);
+		safe_exit(map, MALLOC);
 	}
 	ft_lstadd_back(map_list, new);
 }
@@ -70,11 +53,7 @@ void	get_map(char *line, t_map *map)
 
 	map_list = malloc(sizeof(t_list *));
 	if (!map_list)
-	{
-		ft_free(&line);
-		print_error(1, "Error\nMalloc fail");
-		safe_exit(map, EXIT_FAILURE);
-	}
+		return (ft_free(&line), safe_exit(map, MALLOC));
 	*map_list = NULL;
 	while (line)
 	{
@@ -84,19 +63,27 @@ void	get_map(char *line, t_map *map)
 	}
 	list_to_arr(map_list, map);
 	check_map(map);
+	check_textures(map);
 }
 
 void	set_colors(char **colors, int *color, t_map *map)
 {
-	int	r;
-	int	g;
-	int	b;
+	char	*tmp;
+	int		i;
+	int		r;
+	int		g;
+	int		b;
 
 	if (!colors[0] || !colors[1] || !colors[2] || colors[3])
+		return (free_array(&colors), safe_exit(map, COLOR));
+	i = -1;
+	while (++i <= 2)
 	{
-		free_array(&colors);
-		print_error(1, "Error\nInvalid color");
-		safe_exit(map, EXIT_FAILURE);
+		tmp = ft_strtrim(colors[i], " \t\v\n\f\r");
+		if (!tmp)
+			return (free_array(&colors), safe_exit(map, MALLOC));
+		free(colors[i]);
+		colors[i] = tmp;
 	}
 	r = get_number(colors[0], colors, map);
 	g = get_number(colors[1], colors, map);
@@ -105,38 +92,25 @@ void	set_colors(char **colors, int *color, t_map *map)
 	free_array(&colors);
 }
 
-unsigned char	get_number(char *str, char **colors, t_map *map)
+int	get_number(char *str, char **colors, t_map *map)
 {
-	unsigned char	color_code;
-	char			*trimmed_str;
-	int				num;
-	char			*new_str;
+	int		num;
+	char	*new_str;
 
-	trimmed_str = ft_strtrim(str, " \t\v\n\f\r");
-	if (!trimmed_str)
-	{
-		free_array(&colors);
-		print_error(1, "Error\nMalloc fail");
-		safe_exit(map, EXIT_FAILURE);
-	}
-	num = ft_atoi(trimmed_str);
-	color_code = num;
+	num = ft_atoi(str);
 	new_str = ft_itoa(num);
 	if (!new_str)
 	{
 		free_array(&colors);
-		ft_free(&trimmed_str);
-		print_error(1, "Error\nMalloc fail");
-		safe_exit(map, EXIT_FAILURE);
+		ft_free(&str);
+		safe_exit(map, MALLOC);
 	}
-	if (ft_strncmp(trimmed_str, new_str, ft_strlen(new_str) + 1) != 0
-		|| (int)color_code != num)
+	if (ft_strncmp(str, new_str, ft_strlen(new_str) + 1) != 0
+		|| (int)(unsigned char)num != num)
 	{
 		free_array(&colors);
 		ft_free(&new_str);
-		ft_free(&trimmed_str);
-		print_error(1, "Error\nInvalid color");
-		safe_exit(map, EXIT_FAILURE);
+		safe_exit(map, COLOR);
 	}
-	return (ft_free(&new_str), ft_free(&trimmed_str), color_code);
+	return (ft_free(&new_str), num);
 }
